Nearly working!
Successfully iterates over unread emails (not merely new ones as-they-arrive..),
and runs lua on them, but I overestimated Luar's recursive reflection, so anything
deeper than the first level doesn't seem to get its methods attached or behave as
a lua-type. Meaning, I'm currently having to shadow a bunch of upstream structs
and bolt on some utility functions. This is providing a good opportunity to add
sanitisation and deduplication though, so it's no biggie.

Probably within a commit or two this will start working in a very basic way, then
I start addressing the issue of moderator commands..