Lots of stuff here, whoops.
Closes #2. Thanks to @takching and @korylprince for their help.

Added ability to define a string->string map in config file which is
accessible then in the eventLoop. This may help make generic eventLoops
for various list-types, with the customisation stuff going in the config
files.

Added new parameter to eventLoop function; the config options!

Still not exactly working, but getting there. Will be refactoring some
API then to make things cleaner at the Lua end.