Works! Removed sending-to-self "bug". Added little features.
Closes #1.

Added PollFrequency config option, which is seconds-between-IMAP-polls
when the inbox is emptied. The default for imapclient had been 5m, which
was terrible for a discussion list. The default is now 60 seconds, but I
prefer 30 seconds.

Added ability to exclude email addresses from recipient list at the Go-level,
only used to exclude the "ListAddress" configuration option from recipients
right now, which avoids bounce errors due to cyclical emails (even if,
unbeknownst to the email server, we've added a header to avoid send-loops).