Member-edit mode seems to work. Fixed big bug due to email.Sender being a blacklisted property.
Elaborating: added a mod "sub" to perform the most common management job for a list; adding & removing
subscribers, and editing subscriber details. It works at least partially for adding people but is not
yet extensively tested.

The bug was in eventLoop management; now that methods are whitelisted in Luar, so too are
properties, and email.Sender (a property containing a string) was not yet whitelisted, so
emails were failing to process in a normal eventloop that uses Sender to check for sender
permissions, etcetera.